Edited some information, grammar and punctuation
I just rearranged/replaced some words with others for grammar and made some punctuation corrections (including changing " ` " to " ' ". I also changed Melisa Chan (... graduate student) to (...undergraduate) student, not sure if that matters (since I do hope to be a graduate student in the future) but just in case it does.